Refactoring to include a setup.pu
