+++ core/__init__.py
@@ -1,0 +1,4 @@
+from .network import Network
+from .link import Link, Pipe, Pump, Valve
+from .node import Node, Junction, Reservoir, Tank
+from .objectcollection import ObjectCollection
