+++ core/lazy_property.py
