+++ core/objectcollection.py
