+++ epynet/network.py
@@ -1,432 +1,255 @@
-""" EPYNET Classes """
-import atexit
+import core
+import pandas as pd
+import datetime
+#from tqdm import tqdm
+from time import sleep
+import epynetUtils
 
-from . import epanet2
-from .objectcollection import ObjectCollection
-from .node import Junction, Tank, Reservoir
-from .link import Pipe, Valve, Pump
-from .curve import Curve
-from .pattern import Pattern
-import os
+# TODO: remove global variables
+actuators_update_dict = {}
+# step_count = 0
 
-class Network(object):
-    """ self.epANET Network Simulation Class """
-    def __init__(self, inputfile=None, units=epanet2.EN_CMH, headloss=epanet2.EN_DW, charset='UTF8'):
 
-        # create multithreaded EPANET instance
-        self.ep = epanet2.EPANET2(charset=charset)
+class WaterDistributionNetwork(core.Network):
+    """Class of the network inherited from Epynet.Network"""
+    def __init__(self, inpfile: str):
+        super().__init__(inputfile=inpfile)
+        self.df_nodes_report = None
+        self.df_links_report = None
+        self.times = []
+        # Interactive flag can be set in run() or in init_simulation() if you want to build manually the step-by-step
+        self.interactive = False
+        self.network_state = pd.Series()
 
-        if inputfile:
-            self.inputfile = inputfile
-            self.rptfile = self.inputfile[:-3]+"rpt"
-            self.binfile = self.inputfile[:-3]+"bin"
-            self.ep.ENopen(self.inputfile, self.rptfile, self.binfile)
+    def set_time_params(self, duration=None, hydraulic_step=None, pattern_step=None, report_step=None, start_time=None,
+                        rule_step=None):
+        """
+        Set the time parameters before the simulation (unit: seconds)
+        :param duration: EN_DURATION
+        :param hydraulic_step: EN_HYDSTEP
+        :param pattern_step: EN_PATTERNSTEP
+        :param report_step: EN_REPORTSTEP
+        :param start_time: EN_STARTTIME
+        :param rule_step: EN_RULESTEP
+        """
+        if duration is not None:
+            self.ep.ENsettimeparam(epynetUtils.get_time_param_code('EN_DURATION'), duration)
+        if hydraulic_step is not None:
+            self.ep.ENsettimeparam(epynetUtils.get_time_param_code('EN_HYDSTEP'), hydraulic_step)
+        if pattern_step is not None:
+            self.ep.ENsettimeparam(epynetUtils.get_time_param_code('EN_PATTERNSTEP'), pattern_step)
+        if report_step is not None:
+            self.ep.ENsettimeparam(epynetUtils.get_time_param_code('EN_REPORTSTEP'), report_step)
+        if start_time is not None:
+            self.ep.ENsettimeparam(epynetUtils.get_time_param_code('EN_STARTTIME'), start_time)
+        if rule_step is not None:
+            self.ep.ENsettimeparam(epynetUtils.get_time_param_code('EN_RULESTEP'), rule_step)
+
+    def set_demand_pattern(self, uid: str, values=None, junctions=None):
+        """
+        Set a base-demand pattern for junctions if exists, otherwise it creates and set a new pattern
+        :param uid: pattern id
+        :param values: list of multipliers, None if already existing
+        :param junctions: list of junctions to which we want to set the pattern
+        """
+        if values is None:
+            if uid not in self.patterns.uid:
+                raise KeyError("Chosen pattern id doesn't exist")
         else:
-            self.inputfile = False
-
-            self.rptfile = ""
-            self.binfile = ""
-
-            self.ep.ENinit(self.rptfile.encode(), self.binfile.encode(), units, headloss)
-
-        self.vertices = {}
-        # prepare network data
-        self.nodes = ObjectCollection()
-        self.junctions = ObjectCollection()
-        self.reservoirs = ObjectCollection()
-        self.tanks = ObjectCollection()
-
-        self.links = ObjectCollection()
-        self.pipes = ObjectCollection()
-        self.valves = ObjectCollection()
-        self.pumps = ObjectCollection()
-
-        self.curves = ObjectCollection()
-        self.patterns = ObjectCollection()
-
-        self.solved = False
-        self.solved_for_simtime = None
-
-        self.load_network()
-
-    def load_network(self):
-        """ Load network data """
-        # load nodes
-        for index in range(1, self.ep.ENgetcount(epanet2.EN_NODECOUNT)+1):
-            # get node type
-            node_type = self.ep.ENgetnodetype(index)
-            uid = self.ep.ENgetnodeid(index)
-
-            if node_type == 0:
-                node = Junction(uid, self)
-                self.junctions[node.uid] = node
-            elif node_type == 1:
-                node = Reservoir(uid, self)
-                self.reservoirs[node.uid] = node
-                self.nodes[node.uid] = node
+            if uid in self.patterns.uid:
+                self.patterns[uid].values = values
             else:
-                node = Tank(uid, self)
-                self.tanks[node.uid] = node
+                self.add_pattern(uid, values)
+        if junctions:
+            for junc in junctions:
+                junc.pattern = uid
 
-            self.nodes[node.uid] = node
-
-
-        # load links
-        for index in range(1, self.ep.ENgetcount(epanet2.EN_LINKCOUNT)+1):
-            link_type = self.ep.ENgetlinktype(index)
-            uid = self.ep.ENgetlinkid(index)
-            # pipes
-            if link_type <= 1:
-                link = Pipe(uid, self)
-                self.pipes[link.uid] = link
-            elif link_type == 2:
-                link = Pump(uid, self)
-                self.pumps[link.uid] = link
-            elif link_type >= 3:
-                link = Valve(uid, self)
-                self.valves[link.uid] = link
-
-            self.links[link.uid] = link
-            link_nodes = self.ep.ENgetlinknodes(index)
-            link.from_node = self.nodes[self.ep.ENgetnodeid(link_nodes[0])]
-            link.from_node.links[link.uid] = link
-            link.to_node = self.nodes[self.ep.ENgetnodeid(link_nodes[1])]
-            link.to_node.links[link.uid] = link
-
-
-        # load curves 
-
-        for index in range(1, self.ep.ENgetcount(epanet2.EN_CURVECOUNT)+1):
-            uid = self.ep.ENgetcurveid(index)
-            self.curves[uid] = Curve(uid, self)
-
-        # load patterns
-        for index in range(1, self.ep.ENgetcount(epanet2.EN_PATCOUNT)+1):
-            uid = self.ep.ENgetpatternid(index)
-            self.patterns[uid] = Pattern(uid, self)
-
-    def reset(self):
-
-        self.solved = False
-        self.solved_for_simtime = None
-
-        for link in self.links:
-            link.reset()
-        for node in self.nodes:
-            node.reset()
-
-    def delete_node(self, uid):
-        index = self.ep.ENgetnodeindex(uid)
-        node_type = self.ep.ENgetnodetype(index)
-
-        for link in list(self.nodes[uid].links):
-            self.delete_link(link.uid);
-
-        del self.nodes[uid]
-
-        if node_type == epanet2.EN_JUNCTION:
-            del self.junctions[uid]
-        elif node_type == epanet2.EN_RESERVOIR:
-            del self.reservoirs[uid]
-        elif node_type == epanet2.EN_TANK:
-            del self.tanks[uid]
-
-        self.ep.ENdeletenode(index)
-
-        self.invalidate_nodes()
-        self.invalidate_links()
-
-    def delete_link(self, uid):
-
-        index = self.ep.ENgetlinkindex(uid)
-        link_type = self.ep.ENgetlinktype(index)
-
-        link = self.links[uid]
-        del link.from_node.links[uid]
-        del link.to_node.links[uid]
-
-        del self.links[uid]
-
-        if link_type == epanet2.EN_PIPE or link_type == epanet2.EN_CVPIPE:
-            del self.pipes[uid]
-        elif link_type == epanet2.EN_PUMP:
-            del self.pumps[uid]
+    def demand_model_summary(self):
+        """
+        Print information related to the current demand model
+        """
+        dm_type, pmin, preq, pexp = self.ep.ENgetdemandmodel()
+        if dm_type == 0:
+            print("Running a demand driven analysis...")
         else:
-            del self.valves[uid]
+            print("Running a pressure driven analysis...")
+            print("-> Minimum pressure: {:.2f}".format(pmin))
+            print("-> Required pressure: {:.2f}".format(preq))
+            print("-> Exponential pressure: {:.2f}".format(pexp))
 
-        self.ep.ENdeletelink(index)
-
-        self.invalidate_nodes()
-        self.invalidate_links()
-
-
-    def add_reservoir(self, uid, x, y, elevation=0):
-
-        self.ep.ENaddnode(uid, epanet2.EN_RESERVOIR)
-
-        index = self.ep.ENgetnodeindex(uid)
-
-        self.ep.ENsetcoord(index, x, y)
-
-        node = Reservoir(uid, self)
-        node.elevation = elevation
-
-        self.reservoirs[uid] = node
-        self.nodes[uid] = node
-
-        self.invalidate_nodes()
-
-        return node
-
-    def add_junction(self, uid, x, y, basedemand=0, elevation=0):
-        self.ep.ENaddnode(uid, epanet2.EN_JUNCTION)
-        index = self.ep.ENgetnodeindex(uid)
-        self.ep.ENsetcoord(index, x, y)
-        node = Junction(uid, self)
-        self.junctions[uid] = node
-        self.nodes[uid] = node
-
-        # configure node
-        node.basedemand = basedemand
-        node.elevation = elevation
-
-        self.invalidate_nodes()
-
-        return node
-
-    def add_tank(self, uid, x, y, diameter=0, maxlevel=0, minlevel=0, tanklevel=0):
-        self.ep.ENaddnode(uid, epanet2.EN_TANK)
-        index = self.ep.ENgetnodeindex(uid)
-        self.ep.ENsetcoord(index, x, y)
-        node = Tank(uid, self)
-        self.tanks[uid] = node
-        self.nodes[uid] = node
-        # config tank
-        node.diameter = diameter
-        node.maxlevel = maxlevel
-        node.minlevel = minlevel
-        node.tanklevel = tanklevel
-
-        self.invalidate_nodes()
-
-        return node
-
-    def add_pipe(self, uid, from_node, to_node, diameter=100, length=10, roughness=0.1, check_valve=False):
-
-        from_node = from_node if isinstance(from_node, str) else from_node.uid
-        to_node = to_node if isinstance(to_node, str) else to_node.uid
-
-        if check_valve:
-            self.ep.ENaddlink(uid, epanet2.EN_CVPIPE, from_node, to_node)
+    def run(self, interactive=False, status_dict=None):
+        """
+        Run method wrapper to set the interactivity option (and others in the future related to RL)
+        :param interactive: to update the actuators with own values
+        :param status_dict: dictionary with predefined updates (just to test, it will be removed)
+        TODO: remove status_dict
+        """
+        global actuators_update_dict
+        if status_dict and interactive:
+            actuators_update_dict = status_dict
+            self.interactive = interactive
         else:
-            self.ep.ENaddlink(uid, epanet2.EN_PIPE, from_node, to_node)
+            self.interactive = False
+        self.__run()
 
-        link = Pipe(uid, self)
-
-        link.diameter = diameter
-        link.length = length
-        link.roughness = roughness
-
-        link.from_node = self.nodes[from_node]
-        link.to_node = self.nodes[to_node]
-        link.to_node.links[link.uid] = link
-        link.from_node.links[link.uid] = link
-        self.pipes[uid] = link
-        self.links[uid] = link
-
-        # set link properties
-        link.diameter = diameter
-        link.length = length
-
-        self.invalidate_links()
-
-        return link
-
-    def add_pump(self, uid, from_node, to_node, speed=0):
-
-        from_node = from_node if isinstance(from_node, str) else from_node.uid
-        to_node = to_node if isinstance(to_node, str) else to_node.uid
-
-        self.ep.ENaddlink(uid, epanet2.EN_PUMP, from_node, to_node)
-        link = Pump(uid, self)
-        link.speed = speed
-        link.from_node = self.nodes[from_node]
-        link.speed = speed
-        link.to_node = self.nodes[to_node]
-        link.to_node.links[link.uid] = link
-        link.from_node.links[link.uid] = link
-        self.pumps[uid] = link
-        self.links[uid] = link
-
-        self.invalidate_links()
-
-        return link
-
-    def add_curve(self, uid, values):
-        self.ep.ENaddcurve(uid)
-
-        curve = Curve(uid, self)
-        curve.values = values
-        self.curves[uid] = curve
-
-        return curve
-
-    def add_pattern(self, uid, values):
-        self.ep.ENaddpattern(uid)
-        pattern = Pattern(uid, self)
-        pattern.values = values
-        self.patterns[uid] = pattern
-
-        return pattern
-
-    def add_valve(self, uid, valve_type, from_node, to_node, diameter=100, setting=0):
-
-        from_node = from_node if isinstance(from_node, str) else from_node.uid
-        to_node = to_node if isinstance(to_node, str) else to_node.uid
-
-        if valve_type.lower() == "gpv":
-            valve_type_code = epanet2.EN_GPV
-        elif valve_type.lower() == "fcv":
-            valve_type_code = epanet2.EN_FCV
-        elif valve_type.lower() == "pbv":
-            valve_type_code = epanet2.EN_PBV
-        elif valve_type.lower() == "tcv":
-            valve_type_code = epanet2.EN_TCV
-        elif valve_type.lower() == "prv":
-            valve_type_code = epanet2.EN_PRV
-        elif valve_type.lower() == "psv":
-            valve_type_code = epanet2.EN_PSV
-        else:
-            raise ValueError("Unknown Valve Type")
-
-        self.ep.ENaddlink(uid, valve_type_code, from_node, to_node)
-        link = Valve(uid, self)
-        link.diameter = diameter
-        link.setting = setting
-        link.from_node = self.nodes[from_node]
-        link.to_node = self.nodes[to_node]
-        link.to_node.links[link.uid] = link
-        link.from_node.links[link.uid] = link
-        self.valves[uid] = link
-        self.links[uid] = link
-
-        self.invalidate_links()
-
-        return link
-
-    def invalidate_links(self):
-        # set network as unsolved
-        self.solved = False
-        # reset link index caches
-        for link in self.links:
-            link._index = None
-
-    def invalidate_nodes(self):
-        # set network as unsolved
-        self.solved = False
-        # reset node index caches
-        for node in self.nodes:
-            node._index = None
-
-    def solve(self, simtime=0):
-        """ Solve Hydraulic Network for Single Timestep"""
-        if self.solved and self.solved_for_simtime == simtime:
-            return
-
-        self.reset()
-        self.ep.ENsettimeparam(4, simtime)
-        self.ep.ENopenH()
-        self.ep.ENinitH(0)
-        self.ep.ENrunH()
-        self.ep.ENcloseH()
-        self.solved = True
-        self.solved_for_simtime = simtime
-
-    def run(self):
-        self.reset()
-        self.time = []
-        # open network
-        self.ep.ENopenH()
-        self.ep.ENinitH(0)
-
-        simtime = 0
+    def __run(self):
+        """
+        Step by step simulation: the idea is to put inside this function the online RL algorithm
+        """
+        self.init_simulation(interactive=self.interactive)
+        curr_time = 0
         timestep = 1
 
-        self.solved = True
-
+        #progress_bar = tqdm(total=self.ep.ENgettimeparam(0))
+        # timestep becomes 0 the last hydraulic step
         while timestep > 0:
-            self.ep.ENrunH()
-            timestep = self.ep.ENnextH()
-            self.time.append(simtime)
-            self.load_attributes(simtime)
-            simtime += timestep
+            timestep, state = self.simulate_step(curr_time=curr_time, actuators_status=actuators_update_dict)
+            curr_time += timestep
 
-    def load_attributes(self, simtime):
-        for node in self.nodes:
-            # clear cached values
-            node._values = {}
-            for property_name in node.properties.keys():
-                if property_name not in node.results.keys():
-                    node.results[property_name] = []
-                node.results[property_name].append(node.get_property(node.properties[property_name]))
+            # in DHALSIM here there should be:
+            # update_state(state)
 
-            # check if it's junction to add basedemand with pattern to results
-            if node.node_type == 'Junction':
-                if 'basedemand' not in node.results.keys():
-                    node.results['basedemand'] = []
-                # if pattern not set it takes the basedemand as it is
-                if node.basedemand > 0 and node.pattern.uid is not '1':
-                    pattern_step = self.ep.ENgettimeparam(3)
-                    node.results['basedemand'].append(
-                        node.basedemand * node.pattern.values[(simtime // pattern_step) % len(node.pattern.values)])
-                else:
-                    node.results['basedemand'].append(node.basedemand)
+            # TODO: remember to comment in DHALSIM
+            # if timestep != 0 and self.interactive:
+            #    self.update_actuators_status()
+            sleep(0.01)
+            #progress_bar.update(timestep)
 
-            node.times.append(simtime)
+        #progress_bar.close()
+        self.ep.ENcloseH()
+        self.solved = True
+        self.create_df_reports()
 
-        for link in self.links:
-            # clear cached values
-            link._values = {}
-            for property_name in link.properties.keys():
-                if property_name not in link.results.keys():
-                    link.results[property_name] = []
-                link.results[property_name].append(link.get_property(link.properties[property_name]))
-            link.times.append(simtime)
+    def init_simulation(self, interactive=False):
+        """
+         Initialize the network simulation
+        """
+        self.interactive = interactive
+        self.reset()
+        self.times = []
+        self.ep.ENopenH()
+        self.ep.ENinitH(flag=0)
 
-    def save_inputfile(self, name):
-        self.ep.ENsaveinpfile(name)
+    def simulate_step(self, curr_time, actuators_status=None):
+        """
+        Simulation of one step from the given time
+        :param actuators_status: dictionary with status update for actuators
+        :param curr_time: current simulation time
+        :return: time until the next event, if 0 the simulation is going to end
+        """
+        self.ep.ENrunH()
+        timestep = self.ep.ENnextH()
 
-    def get_vertices(self, link_uid):
-        if self.vertices == {}:
-            self.parse_vertices()
-        return self.vertices.get(link_uid, [])
+        # append new values to reports
+        self.times.append(curr_time)
+        self.load_attributes(curr_time)
 
-    def parse_vertices(self):
-        vertices = False
-        if not self.inputfile or len(self.vertices) > 0:
-            return
+        # update the status of actuators after the first step
+        # TODO: DHALSIM works with the status update here
+        if actuators_status and self.interactive and timestep != 0:
+                self.update_actuators_status(actuators_status)
 
-        with open(self.inputfile, 'rb') as handle:
-            for line in handle.readlines():
-                if b'[VERTICES]' in line:
-                    vertices = True
-                    continue
-                elif b'[' in line:
-                    vertices = False
+        return timestep, self.get_network_state()
 
-                if b";" in line:
-                    continue
+    def update_actuators_status(self, new_status):
+        """
+        Set actuators (pumps and valves) status to a new current state
+        :param new_status: will be used in future with RL
+        TODO: update with new_status and remove step_count
+        """
+        # global step_count
+        for uid in new_status.keys():
+            # self.links[uid].status = new_status[uid][step_count % 2]
+            self.links[uid].status = new_status[uid]
+        # step_count += 1
 
-                if vertices:
-                    components = [c.strip() for c in line.decode(self.ep.charset).split()]
-                    if len(components) < 3:
-                        continue
-                    if components[0] not in self.vertices:
-                        self.vertices[components[0]] = []
-                    self.vertices[components[0]].append((float(components[1]), float(components[2])))
+    def get_network_state(self):
+        """
+        Retrieve the current values of the network in the form a pandas series of dictionaries.
+        The collected values are referred to:
+            - tanks: {pressure}
+            - junctions: {pressure}
+            - pumps: {status, flow}
+            - valves: {status, flow}
+        :return: the series with the above enlisted values
+        """
+        network_state = pd.Series()
+        for uid in self.tanks.results.index.append(self.junctions.results.index):
+            nodes_dict = {key: self.nodes[uid].results[key][-1] for key in ['pressure']}
+            network_state[uid] = nodes_dict
+
+        if self.valves:
+            for uid in self.pumps.results.index.append(self.valves.results.index):
+                links_dict = {key: self.links[uid].results[key][-1] for key in ['status', 'flow']}
+                network_state[uid] = links_dict
+        else:
+            for uid in self.pumps.results.index:
+                links_dict = {key: self.pumps[uid].results[key][-1] for key in ['status', 'flow']}
+                network_state[uid] = links_dict
+        return network_state
+
+    def create_df_reports(self):
+        """
+        Create nodes and links report dataframes - 3 level dataframe
+        How to access: df['node', 'id', 'property'] -> column
+        TODO: create a unique 4 level dataframe with 0 level distinguishing between node and link
+        """
+        if self.df_nodes_report is not None:
+            del self.df_nodes_report
+        if self.df_links_report is not None:
+            del self.df_links_report
+
+        tanks_ids = [uid for uid in self.tanks.uid]
+        junctions_ids = [uid for uid in self.junctions.uid]
+        tanks_iterables = [['tanks'], tanks_ids, ['head', 'pressure']]
+        junct_iterables = [['junctions'], junctions_ids, ['head', 'pressure', 'basedemand', 'actual_demand', 'demand_deficit']]
+        tanks_indices = pd.MultiIndex.from_product(iterables=tanks_iterables, names=["node", "id", "properties"])
+        junctions_indices = pd.MultiIndex.from_product(iterables=junct_iterables, names=["node", "id", "properties"])
+
+        # We use timestamp as index for both nodes and links dataframes
+        times = [datetime.timedelta(seconds=time) for time in self.times]
+
+        # Nodes dataframes creation
+        df_tanks = pd.DataFrame(columns=tanks_indices, index=times)
+        df_junctions = pd.DataFrame(columns=junctions_indices, index=times)
+
+        # Dataframe filling
+        for i, j in zip(df_tanks.columns.get_level_values(1), df_tanks.columns.get_level_values(2)):
+            df_tanks['tanks', i, j] = self.tanks.results[i][j]
+        for i, j in zip(df_junctions.columns.get_level_values(1), df_junctions.columns.get_level_values(2)):
+            df_junctions['junctions', i, j] = self.junctions.results[i][j]
+
+        self.df_nodes_report = pd.concat([df_tanks, df_junctions], axis=1)
+
+        # We can assume that there is always at least one pump in each network, since would be pointless to study a wds
+        # without this kind of links.
+        pumps_ids = [uid for uid in self.pumps.uid]
+        pumps_iterables = [['pumps'], pumps_ids, ['flow', 'energy', 'status']]
+        pumps_indices = pd.MultiIndex.from_product(iterables=pumps_iterables, names=["link", "id", "properties"])
+        df_pumps = pd.DataFrame(columns=pumps_indices, index=times)
+
+        # Pump dataframe filling and columns renaming
+        for i, j in zip(df_pumps.columns.get_level_values(1), df_pumps.columns.get_level_values(2)):
+            df_pumps['pumps', i, j] = self.pumps.results[i][j]
+
+        self.df_links_report = df_pumps
+
+        # We cannot do the same assumption for valves, as we can see in "anytown" network
+        if self.valves:
+            valves_ids = [uid for uid in self.valves.uid]
+            valves_iterables = [['valves'], valves_ids, ['velocity', 'flow', 'status']]
+            valves_indices = pd.MultiIndex.from_product(iterables=valves_iterables, names=["link", "id", "properties"])
+
+            df_valves = pd.DataFrame(columns=valves_indices, index=times)
+
+            # Valves dataframe filling and columns renaming
+            for i, j in zip(df_valves.columns.get_level_values(1), df_valves.columns.get_level_values(2)):
+                df_valves['valves', i, j] = self.valves.results[i][j]
+
+            self.df_links_report = pd.concat([df_pumps, df_valves], axis=1)
 
 
 
-    def close(self):
-        print('closing')
-        self.ep.ENdeleteproject()
+
+
+
